fix: Mavlink enums can have values beyond the range of Int

+++ api/src/main/kotlin/com/urbanmatrix/mavlink/api/MavEnum.kt
@@ -1,5 +1,5 @@
 package com.urbanmatrix.mavlink.api
 
 interface MavEnum {
-    val value: Int
+    val value: Long
 }
